+++ src/Sentry.Unity/Protocol/Unity.cs
@@ -52,13 +52,19 @@ public sealed class Unity : ISentryJsonSerializable
     /// </summary>
     public string? TargetFrameRate { get; set; }
 
+    /// <summary>
+    /// The active scene's name
+    /// </summary>
+    public string? ActiveSceneName { get; set; }
+
     internal Unity Clone()
         => new()
         {
             InstallMode = InstallMode,
             CopyTextureSupport = CopyTextureSupport,
             RenderingThreadingMode = RenderingThreadingMode,
-            TargetFrameRate = TargetFrameRate
+            TargetFrameRate = TargetFrameRate,
+            ActiveSceneName = ActiveSceneName
         };
 
     public void WriteTo(Utf8JsonWriter writer, IDiagnosticLogger? logger)
@@ -92,6 +98,11 @@ public sealed class Unity : ISentryJsonSerializable
             writer.WriteString("target_frame_rate", TargetFrameRate);
         }
 
+        if (!string.IsNullOrWhiteSpace(ActiveSceneName))
+        {
+            writer.WriteString("active_scene_name", ActiveSceneName);
+        }
+
         writer.WriteEndObject();
     }
 
@@ -102,7 +113,8 @@ public sealed class Unity : ISentryJsonSerializable
             InstallMode = json.GetPropertyOrNull("install_mode")?.GetString(),
             CopyTextureSupport = json.GetPropertyOrNull("copy_texture_support")?.GetString(),
             RenderingThreadingMode = json.GetPropertyOrNull("rendering_threading_mode")?.GetString(),
-            TargetFrameRate = json.GetPropertyOrNull("target_frame_rate")?.GetString()
+            TargetFrameRate = json.GetPropertyOrNull("target_frame_rate")?.GetString(),
+            ActiveSceneName = json.GetPropertyOrNull("active_scene_name")?.GetString()
         };
 
     public string ToJsonString(IDiagnosticLogger? logger = null)